Tray icon added
New Update adds a tray icon that you can enable or disable easily. this way you can hide the app and use it and you can open it  or close any time by left clicking  on the tray icon or right clicking and selecting the option in the context menu.